requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo;python_version<"3.9"
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==41.0.1
Deprecated==1.2.13
Django==4.0.5
django-auto-logout==0.5.0
django-celery-results==2.5.1
django-crispy-forms==1.12.0
django-environ==0.9.0
django-extensions==3.2.3
django-force-logout==3.2.2
idna==3.4
kombu==5.3.1
MarkupSafe==2.1.3
ms-identity-web==0.0.4
msal==1.22.0
packaging==21.3
Pillow==9.4.0
prompt-toolkit==3.0.30
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.3.6
wrapt==1.14.1
python-dotenv<0.22